refactor(app): hoist Bootstrap and BuildServices out of cmd/arcade
`buildServices` and the dependency wiring around it lived as private
helpers inside `cmd/arcade/main.go`, which meant test harnesses that
want to boot arcade in-process either had to launch the binary or
duplicate every kafka/store/teranode/merkle-client setup line.

Moves both into a new top-level `app` package with two entry points:
`app.Bootstrap(ctx, cfg, logger)` returns a `*Deps` plus a cleanup
func (closing publisher → teranode client → store → producer in
reverse order), and `app.BuildServices(deps)` returns the slice of
services to run for the configured mode. `cmd/arcade/main.go`
becomes a thin supervisor that calls the two and handles signals.

No behavior change. The full default test suite passes; this lands
ahead of the e2e harness which will reuse the same boot path.
